Choices moved to apropriate folder.
