+++ src/Pure.Primitives.Choices/Guid/GuidChoice.cs
@@ -1,5 +1,5 @@
-﻿using Pure.Primitives.Bool;
-using System;
+﻿using Pure.Primitives.Abstractions.Bool;
+using Pure.Primitives.Abstractions.Guid;
 
 namespace Pure.Primitives.Guid;
